src/requirements.txt
accelerate==0.31.0
datasets==2.20.0
huggingface_hub==0.23.4
lightning==2.3.3
numpy==2.1.0
pandas==2.2.2
Pillow==10.4.0
psutil==6.0.0
pyarrow==16.1.0
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
spacy==3.7.6
torch==2.4.0+cu124
torch_xla==2.4.0
tqdm==4.66.4
transformers==4.43.3
trl==0.9.6
